restructure pw polynomial control example
